requirements.txt
openai-whisper
pyaudio
pyttsx3
keyboard
python-dotenv
requests
ollama
numpy
torch
torchaudio